+++ src/C3D/Extensions/Aspire/IISExpress/Resources/IISExpressSiteResource.cs
@@ -1,0 +1,10 @@
+﻿using Aspire.Hosting;
+using Aspire.Hosting.ApplicationModel;
+
+namespace C3D.Extensions.Aspire.IISExpress.Resources;
+
+public class IISExpressSiteResource(IISExpressResource iisExpress, string siteName, string workingDirectory)
+    : ExecutableResource(siteName, iisExpress.Path, workingDirectory), IResourceWithServiceDiscovery, IResourceWithEnvironment
+{
+    public IISExpressResource IISExpress => iisExpress;
+}
